refactor!: canonicaldecay -> canonical_decay
