+++ expertsystem/amplitude/canonical_decay.py
