+++ tests/channels/test_y_to_d0d0barpi0pi0.py
@@ -6,7 +6,7 @@ import logging
 
 import pytest
 
-from expertsystem.amplitude.canonicaldecay import CanonicalAmplitudeGenerator
+from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
 from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
 from expertsystem.state.particle import (
     InteractionQuantumNumberNames,
